+++ iam-policy-glue.tf
@@ -5,9 +5,9 @@
  */
 
 resource "aws_iam_role_policy" "glue_for_ecs_task_readwrite" {
-  count = "${ var.enable_gluesync == "" ? 0 : 1 }"
+  count = "${var.enable_gluesync == "" ? 0 : 1}"
   name  = "glue"
-  role  = "${aws_iam_role.apiary_task_readwrite.id}"
+  role  = "${aws_iam_role.apiary_hms_readwrite.id}"
 
   policy = <<EOF
 {