+++ iam-policy-s3-buckets.tf
@@ -5,9 +5,9 @@
  */
 
 resource "aws_iam_role_policy" "s3_data_for_ecs_task_readwrite" {
-  count = "${ length(var.apiary_managed_schemas) == 0 ? 0 : 1 }"
+  count = "${length(var.apiary_managed_schemas) == 0 ? 0 : 1}"
   name  = "s3"
-  role  = "${aws_iam_role.apiary_task_readwrite.id}"
+  role  = "${aws_iam_role.apiary_hms_readwrite.id}"
 
   policy = <<EOF
 {
@@ -30,8 +30,8 @@ resource "aws_iam_role_policy" "s3_data_for_ecs_task_readwrite" {
                               "s3:PutObjectVersionTagging"
                             ],
                   "Resource": [
-                                "${join("\",\"", formatlist("arn:aws:s3:::%s",local.apiary_data_buckets))}",
-                                "${join("\",\"", formatlist("arn:aws:s3:::%s/*",local.apiary_data_buckets))}"
+                                "${join("\",\"", formatlist("arn:aws:s3:::%s", local.apiary_data_buckets))}",
+                                "${join("\",\"", formatlist("arn:aws:s3:::%s/*", local.apiary_data_buckets))}"
                               ]
                 }
               ]
@@ -40,9 +40,9 @@ EOF
 }
 
 resource "aws_iam_role_policy" "s3_data_for_ecs_task_readonly" {
-  count = "${ length(var.apiary_managed_schemas) == 0 ? 0 : 1 }"
+  count = "${length(var.apiary_managed_schemas) == 0 ? 0 : 1}"
   name  = "s3"
-  role  = "${aws_iam_role.apiary_task_readonly.id}"
+  role  = "${aws_iam_role.apiary_hms_readonly.id}"
 
   policy = <<EOF
 {
@@ -55,8 +55,8 @@ resource "aws_iam_role_policy" "s3_data_for_ecs_task_readonly" {
                               "s3:List*"
                             ],
                   "Resource": [
-                                "${join("\",\"", formatlist("arn:aws:s3:::%s",local.apiary_data_buckets))}",
-                                "${join("\",\"", formatlist("arn:aws:s3:::%s/*",local.apiary_data_buckets))}"
+                                "${join("\",\"", formatlist("arn:aws:s3:::%s", local.apiary_data_buckets))}",
+                                "${join("\",\"", formatlist("arn:aws:s3:::%s/*", local.apiary_data_buckets))}"
                               ]
                 }
               ]
@@ -65,9 +65,9 @@ EOF
 }
 
 resource "aws_iam_role_policy" "external_s3_data_for_ecs_task_readwrite" {
-  count = "${ length(var.external_data_buckets) == 0 ? 0 : 1 }"
+  count = "${length(var.external_data_buckets) == 0 ? 0 : 1}"
   name  = "external-s3"
-  role  = "${aws_iam_role.apiary_task_readwrite.id}"
+  role  = "${aws_iam_role.apiary_hms_readwrite.id}"
 
   policy = <<EOF
 {
@@ -90,8 +90,8 @@ resource "aws_iam_role_policy" "external_s3_data_for_ecs_task_readwrite" {
                               "s3:PutObjectVersionTagging"
                             ],
                   "Resource": [
-                                "${join("\",\"", formatlist("arn:aws:s3:::%s",var.external_data_buckets))}",
-                                "${join("\",\"", formatlist("arn:aws:s3:::%s/*",var.external_data_buckets))}"
+                                "${join("\",\"", formatlist("arn:aws:s3:::%s", var.external_data_buckets))}",
+                                "${join("\",\"", formatlist("arn:aws:s3:::%s/*", var.external_data_buckets))}"
                               ]
                 }
               ]
@@ -100,9 +100,9 @@ EOF
 }
 
 resource "aws_iam_role_policy" "external_s3_data_for_ecs_task_readonly" {
-  count = "${ length(var.external_data_buckets) == 0 ? 0 : 1 }"
+  count = "${length(var.external_data_buckets) == 0 ? 0 : 1}"
   name  = "external-s3"
-  role  = "${aws_iam_role.apiary_task_readonly.id}"
+  role  = "${aws_iam_role.apiary_hms_readonly.id}"
 
   policy = <<EOF
 {
@@ -115,8 +115,8 @@ resource "aws_iam_role_policy" "external_s3_data_for_ecs_task_readonly" {
                               "s3:List*"
                             ],
                   "Resource": [
-                                "${join("\",\"", formatlist("arn:aws:s3:::%s",var.external_data_buckets))}",
-                                "${join("\",\"", formatlist("arn:aws:s3:::%s/*",var.external_data_buckets))}"
+                                "${join("\",\"", formatlist("arn:aws:s3:::%s", var.external_data_buckets))}",
+                                "${join("\",\"", formatlist("arn:aws:s3:::%s/*", var.external_data_buckets))}"
                               ]
                 }
               ]