+++ iam-policy-secretsmanager.tf
@@ -6,7 +6,7 @@
 
 resource "aws_iam_role_policy" "secretsmanager_for_ecs_readonly" {
   name = "secretsmanager"
-  role = "${aws_iam_role.apiary_task_readonly.id}"
+  role = "${aws_iam_role.apiary_hms_readonly.id}"
 
   policy = <<EOF
 {
@@ -22,7 +22,7 @@ EOF
 
 resource "aws_iam_role_policy" "secretsmanager_for_ecs_task_readwrite" {
   name = "secretsmanager"
-  role = "${aws_iam_role.apiary_task_readwrite.id}"
+  role = "${aws_iam_role.apiary_hms_readwrite.id}"
 
   policy = <<EOF
 {